+++ app/src/main/java/com/ufrstgi/imr/application/database/ColisManager.java
@@ -114,23 +114,23 @@ public class ColisManager {
         return db.delete(TABLE_NAME, where, whereArgs);
     }
 
-    public Colis getColis(Long id) {
+    public Colis getColis(int id) {
         // Retourne le niveau dont l'id est passé en paramètre
 
         Colis co = new Colis(0,0,0,0,0,0,0,0,0,0);
 
         Cursor c = db.rawQuery("SELECT * FROM "+TABLE_NAME+" WHERE "+KEY_ID_COLIS+"="+id, null);
         if (c.moveToFirst()) {
-            co.setId_colis(c.getLong(c.getColumnIndex(KEY_ID_COLIS)));
+            co.setId_colis(c.getInt(c.getColumnIndex(KEY_ID_COLIS)));
             co.setPoids_colis(c.getFloat(c.getColumnIndex(KEY_POIDS_COLIS)));
             co.setVolume_colis(c.getFloat(c.getColumnIndex(KEY_VOLUME_COLIS)));
             co.setNiveau_batterie_colis(c.getFloat(c.getColumnIndex(KEY_NIVEAU_BATTERIE_COLIS)));
             co.setTemperature_colis(c.getFloat(c.getColumnIndex(KEY_TEMPERATURE_COLIS)));
             co.setCapacite_choc_colis(c.getFloat(c.getColumnIndex(KEY_CAPACITE_CHOC_COLIS)));
-            co.setId_niveau(c.getLong(c.getColumnIndex(KEY_ID_NIVEAU)));
-            co.setId_operation(c.getLong(c.getColumnIndex(KEY_ID_OPERATION)));
-            co.setId_tournee(c.getLong(c.getColumnIndex(KEY_ID_TOURNEE)));
-            co.setId_client(c.getLong(c.getColumnIndex(KEY_ID_CLIENT)));
+            co.setId_niveau(c.getInt(c.getColumnIndex(KEY_ID_NIVEAU)));
+            co.setId_operation(c.getInt(c.getColumnIndex(KEY_ID_OPERATION)));
+            co.setId_tournee(c.getInt(c.getColumnIndex(KEY_ID_TOURNEE)));
+            co.setId_client(c.getInt(c.getColumnIndex(KEY_ID_CLIENT)));
 
             c.close();
         }
